added all tests to MDUserStorage_Tests

+++ MyDictionary_App_SwiftTests/User/MDUserStorage_Tests.swift
@@ -45,14 +45,18 @@ extension MDUserStorage_Tests {
         let storageType: MDStorageType = .memory
         
         userStorage.createUser(Constants_For_Tests.mockedUser, storageType: storageType) { [unowned self] results in
+            
             switch results.first!.result {
+            
             case .success(let createdUser):
+                
                 XCTAssertTrue(createdUser.userId == Constants_For_Tests.mockedUser.userId)
                 XCTAssertTrue(createdUser.nickname == Constants_For_Tests.mockedUser.nickname)
                 XCTAssertTrue(createdUser.password == Constants_For_Tests.mockedUser.password)
                 XCTAssertTrue(createdUser.createdAt == Constants_For_Tests.mockedUser.createdAt)
                 XCTAssertTrue(createdUser.updatedAt == Constants_For_Tests.mockedUser.updatedAt)
                 expectation.fulfill()
+                
             case .failure:
                 XCTExpectFailure()
                 expectation.fulfill()
@@ -69,17 +73,24 @@ extension MDUserStorage_Tests {
         let storageType: MDStorageType = .memory
         
         userStorage.createUser(Constants_For_Tests.mockedUser, storageType: storageType) { [unowned self] createResults in
+            
             switch createResults.first!.result {
+            
             case .success(let createdUser):
+                
                 userStorage.readUser(fromUserID: createdUser.userId, storageType: storageType) { [unowned self] readResults in
+                    
                     switch readResults.first!.result {
+                    
                     case .success(let readUser):
+                        
                         XCTAssertTrue(createdUser.userId == readUser.userId)
                         XCTAssertTrue(createdUser.nickname == readUser.nickname)
                         XCTAssertTrue(createdUser.password == readUser.password)
                         XCTAssertTrue(createdUser.createdAt == readUser.createdAt)
                         XCTAssertTrue(createdUser.updatedAt == readUser.updatedAt)
                         expectation.fulfill()
+                        
                     case .failure:
                         XCTExpectFailure()
                         expectation.fulfill()
@@ -101,17 +112,28 @@ extension MDUserStorage_Tests {
         let storageType: MDStorageType = .memory
         
         userStorage.createUser(Constants_For_Tests.mockedUser, storageType: storageType) { [unowned self] createResults in
+            
             switch createResults.first!.result {
+            
             case .success(let createdUser):
-                self.userStorage.deleteUser(createdUser, storageType: storageType) { deleteResults in
+                
+                self.userStorage.deleteUser(createdUser, storageType: storageType) { [unowned self] deleteResults in
+                    
                     switch deleteResults.first!.result {
+                    
                     case .success(let deleteUser):
+                        
                         XCTAssertTrue(createdUser.userId == deleteUser.userId)
+                        
                         self.userStorage.entitiesIsEmpty(storageType: storageType) { [unowned self] entitiesIsEmptyResults in
+                            
                             switch entitiesIsEmptyResults.first!.result {
+                            
                             case .success(let entitiesIsEmpty):
+                                
                                 XCTAssertTrue(entitiesIsEmpty)
                                 expectation.fulfill()
+                                
                             case .failure:
                                 XCTExpectFailure()
                                 expectation.fulfill()
@@ -143,14 +165,18 @@ extension MDUserStorage_Tests {
         let storageType: MDStorageType = .coreData
         
         userStorage.createUser(Constants_For_Tests.mockedUser, storageType: storageType) { [unowned self] results in
+            
             switch results.first!.result {
+            
             case .success(let createdUser):
+                
                 XCTAssertTrue(createdUser.userId == Constants_For_Tests.mockedUser.userId)
                 XCTAssertTrue(createdUser.nickname == Constants_For_Tests.mockedUser.nickname)
                 XCTAssertTrue(createdUser.password == Constants_For_Tests.mockedUser.password)
                 XCTAssertTrue(createdUser.createdAt == Constants_For_Tests.mockedUser.createdAt)
                 XCTAssertTrue(createdUser.updatedAt == Constants_For_Tests.mockedUser.updatedAt)
                 expectation.fulfill()
+                
             case .failure:
                 XCTExpectFailure()
                 expectation.fulfill()
@@ -167,22 +193,30 @@ extension MDUserStorage_Tests {
         let storageType: MDStorageType = .coreData
         
         userStorage.createUser(Constants_For_Tests.mockedUser, storageType: storageType) { [unowned self] createResults in
+            
             switch createResults.first!.result {
+            
             case .success(let createdUser):
+                
                 userStorage.readUser(fromUserID: createdUser.userId, storageType: storageType) { [unowned self] readResults in
+                    
                     switch readResults.first!.result {
+                    
                     case .success(let readUser):
+                        
                         XCTAssertTrue(createdUser.userId == readUser.userId)
                         XCTAssertTrue(createdUser.nickname == readUser.nickname)
                         XCTAssertTrue(createdUser.password == readUser.password)
                         XCTAssertTrue(createdUser.createdAt == readUser.createdAt)
                         XCTAssertTrue(createdUser.updatedAt == readUser.updatedAt)
                         expectation.fulfill()
+                        
                     case .failure:
                         XCTExpectFailure()
                         expectation.fulfill()
                     }
                 }
+                
             case .failure:
                 XCTExpectFailure()
                 expectation.fulfill()
@@ -199,25 +233,192 @@ extension MDUserStorage_Tests {
         let storageType: MDStorageType = .coreData
         
         userStorage.createUser(Constants_For_Tests.mockedUser, storageType: storageType) { [unowned self] createResults in
+            
             switch createResults.first!.result {
+            
             case .success(let createdUser):
-                self.userStorage.deleteUser(createdUser, storageType: storageType) { deleteResults in
+                
+                self.userStorage.deleteUser(createdUser, storageType: storageType) { [unowned self] deleteResults in
+                    
                     switch deleteResults.first!.result {
+                    
                     case .success(let deleteUser):
+                        
                         XCTAssertTrue(createdUser.userId == deleteUser.userId)
+                        
                         self.userStorage.entitiesIsEmpty(storageType: storageType) { [unowned self] entitiesIsEmptyResults in
+                            
                             switch entitiesIsEmptyResults.first!.result {
+                            
                             case .success(let entitiesIsEmpty):
+                                
                                 XCTAssertTrue(entitiesIsEmpty)
                                 expectation.fulfill()
+                                
                             case .failure:
                                 XCTExpectFailure()
                                 expectation.fulfill()
                             }
+                            
                         }
+                        
                     case .failure:
                         XCTExpectFailure()
                         expectation.fulfill()
+                    }
+                }
+            case .failure:
+                XCTExpectFailure()
+                expectation.fulfill()
+            }
+        }
+        
+        wait(for: [expectation], timeout: Constants_For_Tests.testExpectationTimeout)
+        
+    }
+    
+}
+
+// MARK: - All CRUD
+extension MDUserStorage_Tests {
+    
+    func test_Create_User_In_All_Functionality() {
+        
+        let expectation = XCTestExpectation(description: "Create User In All Expectation")
+        let storageType: MDStorageType = .all
+        
+        var resultCount: Int = .zero
+        
+        userStorage.createUser(Constants_For_Tests.mockedUser, storageType: storageType) { [unowned self] createResults in
+            
+            createResults.forEach { createResult in
+                
+                switch createResult.result {
+                
+                case .success(let createdUser):
+                    
+                    resultCount += 1
+                    
+                    XCTAssertTrue(createdUser.userId == Constants_For_Tests.mockedUser.userId)
+                    XCTAssertTrue(createdUser.nickname == Constants_For_Tests.mockedUser.nickname)
+                    XCTAssertTrue(createdUser.password == Constants_For_Tests.mockedUser.password)
+                    XCTAssertTrue(createdUser.createdAt == Constants_For_Tests.mockedUser.createdAt)
+                    XCTAssertTrue(createdUser.updatedAt == Constants_For_Tests.mockedUser.updatedAt)
+                    
+                    if (resultCount == createResults.count) {
+                        expectation.fulfill()
+                    }
+                    
+                case .failure:
+                    XCTExpectFailure()
+                    expectation.fulfill()
+                }
+                
+            }
+        }
+        
+        wait(for: [expectation], timeout: Constants_For_Tests.testExpectationTimeout)
+        
+    }
+    
+    func test_Read_User_From_All_Functionality() {
+        
+        let expectation = XCTestExpectation(description: "Read User From All Expectation")
+        let storageType: MDStorageType = .all
+        
+        var resultCount: Int = .zero
+        
+        userStorage.createUser(Constants_For_Tests.mockedUser, storageType: storageType) { [unowned self] createResults in
+            
+            switch createResults.first!.result {
+            
+            case .success(let createdUser):
+                
+                userStorage.readUser(fromUserID: createdUser.userId, storageType: storageType) { [unowned self] readResults in
+                    
+                    readResults.forEach { readResult in
+                        
+                        switch readResult.result {
+                        
+                        case .success(let readUser):
+                            
+                            resultCount += 1
+                            
+                            XCTAssertTrue(createdUser.userId == readUser.userId)
+                            XCTAssertTrue(createdUser.nickname == readUser.nickname)
+                            XCTAssertTrue(createdUser.password == readUser.password)
+                            XCTAssertTrue(createdUser.createdAt == readUser.createdAt)
+                            XCTAssertTrue(createdUser.updatedAt == readUser.updatedAt)
+                            
+                            if (resultCount == readResults.count) {
+                                expectation.fulfill()
+                            }
+                            
+                        case .failure:
+                            XCTExpectFailure()
+                            expectation.fulfill()
+                        }
+                    }
+                }
+                
+            case .failure:
+                XCTExpectFailure()
+                expectation.fulfill()
+            }
+            
+        }
+        
+        wait(for: [expectation], timeout: Constants_For_Tests.testExpectationTimeout)
+        
+    }
+    
+    func test_Delete_User_From_All_Functionality() {
+        
+        let expectation = XCTestExpectation(description: "Delete User From All Expectation")
+        let storageType: MDStorageType = .all
+        
+        var resultCount: Int = .zero
+        
+        userStorage.createUser(Constants_For_Tests.mockedUser, storageType: storageType) { [unowned self] createResults in
+            
+            switch createResults.first!.result {
+            
+            case .success(let createdUser):
+                
+                self.userStorage.deleteUser(createdUser, storageType: storageType) { [unowned self] deleteResults in
+                    
+                    deleteResults.forEach { deleteResult in
+                        
+                        switch deleteResult.result {
+                        
+                        case .success(let deleteUser):
+                            
+                            XCTAssertTrue(createdUser.userId == deleteUser.userId)
+                            
+                            self.userStorage.entitiesIsEmpty(storageType: deleteResult.storageType) { [unowned self] entitiesIsEmptyResults in
+                                
+                                switch entitiesIsEmptyResults.first!.result {
+                                
+                                case .success(let entitiesIsEmpty):
+                                    
+                                    resultCount += 1
+                                    
+                                    XCTAssertTrue(entitiesIsEmpty)
+                                    
+                                    if (resultCount == deleteResults.count) {
+                                        expectation.fulfill()
+                                    }
+                                    
+                                case .failure:
+                                    XCTExpectFailure()
+                                    expectation.fulfill()
+                                }
+                            }
+                            
+                        case .failure:
+                            XCTExpectFailure()
+                            expectation.fulfill()
+                        }
                     }
                 }
             case .failure:
